Demian Resource png resave done

+++ GameEngine/GameEngineTexture.h
@@ -26,7 +26,8 @@ public:
 
 	inline float4 GetTextureSize()
 	{
-		return float4(TextureDesc_.Width, TextureDesc_.Height);
+		return float4(static_cast<float>(TextureDesc_.Width),
+			static_cast<float>(TextureDesc_.Height));
 	}
 
 protected:
